Allow DIY to compile under ROCm HIP compiler
The ROCm HIP compiler ( a clang based compiler ) has issues
compiling DIY via VTK-m. In general the issue is a compiler
bug where it incorrectly presumes that the `diy/master`
constructor is being invoked from a device side function.

This is a problem as HIP doesn't support allocations or
frees ( so no keyword new / delete ) in device side functions.
We work around the compiler bug by making the allocation happen in the
definition instead of the declaration.